overhaul to input logic; changed to enums
I saw that a lot of the information was incorrect in the last version, but when I went to go fix it, I found many sources contradicting each other on what the correct property was. I thought this would happen, as chemistry based on observations of the real world, and observations may differ. I ended up sticking to one source to help fill out all of the information.